Add files via upload
This is the source code of my calculator project. As for the planned beta version availability, this source code only contains 50% of the overall progression. The calculator app only works for onshore pipe support spacing calculator and road crossing assessment for now. For both onshore and offshore pipeline wall thickness calculation is still an ongoing development.